gerenciamento_consultas/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.22
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
jinja2>=3.0.1
python-multipart>=0.0.5
email-validator>=2.0.0
bcrypt>=4.2.1
